+++ src/sbs/floor.cpp
@@ -798,7 +798,7 @@ void Floor::AddFillerWalls(const char *texture, float thickness, float CenterX, 
 	sbs->ResetWalls();
 }
 
-Object* Floor::AddSound(const char *name, const char *filename, Ogre::Vector3 position, bool loop, int volume, int speed, float min_distance, float max_distance, float doppler_level, float cone_inside_angle, float cone_outside_angle, float cone_outside_volume, Ogre::Vector3 direction)
+Object* Floor::AddSound(const char *name, const char *filename, Ogre::Vector3 position, bool loop, float volume, int speed, float min_distance, float max_distance, float doppler_level, float cone_inside_angle, float cone_outside_angle, float cone_outside_volume, Ogre::Vector3 direction)
 {
 	//create a looping sound object
 	Sound *sound = new Sound(object, name, false);
